address comments and update with master
