+++ src/datasets/datasets.module.ts
@@ -24,7 +24,6 @@ import { ConfigModule, ConfigService } from "@nestjs/config";
 import { applyHistoryPluginOnce } from "src/common/mongoose/plugins/history.plugin.util";
 import { ProposalsModule } from "src/proposals/proposals.module";
 import { HistoryModule } from "src/history/history.module";
-import { ScientificMetadataValidator } from "src/datasets/utils/scientificMetadata";
 
 @Module({
   imports: [
@@ -99,11 +98,6 @@ import { ScientificMetadataValidator } from "src/datasets/utils/scientificMetada
     DatasetsController,
     DatasetsV4Controller,
   ],
-  providers: [
-    DatasetsService,
-    DatasetsAccessService,
-    DatasetsV4Controller,
-    ScientificMetadataValidator,
-  ],
+  providers: [DatasetsService, DatasetsAccessService, DatasetsV4Controller],
 })
 export class DatasetsModule {}
